Updated links and the requirements section.
Changed all the external links to be free of language selection, so that the respective sites can redirect you to the appropriate page for your language rather than defaulting to English (or Polish in the case of the framework link). Also AMD has changed their support location from support.amd.com to amd.com/support, so that's been updated too.

Rewrote, reformatted and renamed the software section, and included a link to the 2015 Redistributable, since the page recommends you include it with games targeting Windows. If nothing else it'll save people a bit of time googling.

The visual studio section was shortened to include only 2017 and 2019 since that's what the extension supports (also 2013 and 2015 is woefully out of date. 2015 don't even support C# 7.0 without weird hacks), and also specified that the community edition is what people should be downloading, since the linked page also offers VS professional and enterprise as "free trials".

Minor change to drivers section too, no need for "helper links" since the heading is clear enough already.